+++ wiki/Changelog.md
@@ -15,6 +15,7 @@ Window handling
   * Suppressing Windows-caused side effects of "resizing" to current size (#629).
   * Tweaked window-raising to top (previous #652) to not stick on top (mintty/wsltty#47).
   * Reverted capturing of Shift+Ctrl+0 (~#233) to enable language switching (#663).
+  * Capture key after leaving system menu with ESC; insert it and prevent beep.
 
 Configuration
   * Reverted to stripping CR from most configuration strings (mintty/wsltty#46).